Update potentialRelays when entering flight scene
The Relay.potentialRelays list was being built just once, in the class's
static constructor, with a subset of the Vessel objects (those with
comms) that were available at that time.  This was problematic for two
reasons: first, it wouldn't notice changes that occur later (e.g. new
vessels with comms being launched), and second, each entry to the flight
scene seems to create a new set of Vessel objects, so the GetRelayVessel
wouldn't find anything matching the new objects because the relays in
the list would still hold references to the old ones.

Rebuilding the potentialRelays list at each entry to the flight scene
keeps it up-to-date with the current set of vessels and their Vessel
objects.