Fully update the ship list when entering the tracking station
Although the AHShipListListener calls ParseFlyingVessel when the player
returns to the space center, that event fires while the vessel-recovery
window is still open and before the recovered vessel has actually been
removed from FlightGlobals.Vessels, so listFlyingVessel will include the
recovered vessel in that case.  If the player then goes to the tracking
station, AHTrackingStation.CreateMarkers would throw exceptions trying
to create a marker for it, since it wouldn't be found in
FlightGlobals.Vessels anymore.  Updating listFlyingVessel on entry to
the tracking station avoids that.