Store GameObjects, not AHMapMarkers, in AHFlight
When leaving the scene, its GameObjects get destroyed, in no particular
order.  Sometimes it happens that an AHMapMarker's parent GameObject
gets destroyed before the Vessel that the marker is associated with.
Destroying the parent GameObject also destroys the AHMapMarker
component, so if AHFlight.VesselDestroy later tries to access the
component's gameObject property, it throws a NullReferenceException
because the component is no longer associated with a GameObject.

To avoid that, AHFlight now stores references to the actual GameObjects
that it'll need to destroy later, instead of to the AHMapMarker
components attached to them.